+++ Modules/Ruleset.ProgramInfo/Ruleset.ProgramInfo.psm1
@@ -55,6 +55,7 @@ if ($ListPreference)
 #
 
 $ScriptsToProcess = @(
+	"AppxModule"
 	"TargetProgram"
 )