+++ Modules/Ruleset.ProgramInfo/Scripts/AppxModule.ps1
@@ -29,7 +29,8 @@ SOFTWARE.
 # Appx module must be imported in compatibility mode for PowerShell version 7.1+
 if ($PSVersionTable.PSVersion -ge "7.1")
 {
-	# NOTE: New-PSSession for compatibility module is first time executed either here or during Initialize-Connection,
-	# if it fails it's either Windows.PowerShell PS session configuration or WinRM loopback misconfiguration
+	Write-Verbose -Message "[AppxModule.ps1] Importing Appx module" -Verbose
+	# NOTE: New-PSSession for compatibility module is first time executed here and for remote
+	# session also in SessionStartupScript in remote session (Remote module)
 	Import-WinModule -Name Appx -ErrorAction Stop
 }